% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{have_dependencies}
\alias{have_dependencies}
\title{Check whether python and dependencies are available on system}
\usage{
have_dependencies()
}
\value{
boolean TRUE/ FALSE
}
\description{
Check whether python and dependencies are available on system
}
