% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_RFmodel}
\alias{import_RFmodel}
\title{Make RFmodel available for direct python calls}
\usage{
import_RFmodel()
}
\value{
attach python RandomForestClassifier to variable \code{RFmodel}
}
\description{
Convenience wrapper to make the python random forest model for snow layer instability 'p_unstable' accessible in R session
}
\seealso{
\link{computePunstable}
}
