% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarp.snowprofile.pyface-package.R
\docType{package}
\name{sarp.snowprofile.pyface-package}
\alias{sarp.snowprofile.pyface}
\alias{sarp.snowprofile.pyface-package}
\title{sarp.snowprofile.pyface: 'python' Modules from Snowpack and Avalanche Research}
\description{
The development of post-processing functionality for simulated snow profiles by the snow and avalanche community is often done in 'python'. This package aims to make these tools accessible to 'R' users. Currently integrated modules contain functions to calculate dry snow layer instabilities in support of avalache hazard assessments following the publications of Richter, Schweizer, Rotach, and Van Herwijnen (2019) <doi:10.5194/tc-13-3353-2019>, and Mayer, Van Herwijnen, Techel, and Schweizer (2022) <doi:10.5194/tc-2022-34>.
}
\seealso{
\link{computeCritCutLength}, \link{computePunstable}
}
\author{
\strong{Maintainer}: Florian Herla \email{fherla@sfu.ca}

Authors:
\itemize{
  \item Stephanie Mayer \email{stephanie.mayer@slf.ch}
}

Other contributors:
\itemize{
  \item SFU Avalanche Research Program [funder]
}

}
\keyword{internal}
